backend/vietqa-demo/demo-chatbot/requirements.txt
fastapi
uvicorn
requests
qdrant_client
sentence_transformers
langchain_text_splitters
pydantic
openai==0.28
python-multipart 
streamlit
python-dotenv   
opencv-python-headless
PyMuPDF==1.19.0
opencv-python
pdfplumber
pytesseract
minio
langdetect